+++ src/Client/dotnet-graphql/DefaultConsoleOutputActivity.cs
@@ -9,13 +9,15 @@ namespace StrawberryShake.Tools
     {
         private readonly string _activityText;
         private readonly string? _path;
+        private readonly Action _errorReceived;
         private readonly Stopwatch _stopwatch;
         private bool _hasErrors;
 
-        public DefaultConsoleOutputActivity(string activityText, string? path)
+        public DefaultConsoleOutputActivity(string activityText, string? path, Action errorReceived)
         {
             _activityText = activityText;
             _path = path;
+            _errorReceived = errorReceived;
             _stopwatch = Stopwatch.StartNew();
             Console.WriteLine($"{activityText} started.");
         }
@@ -24,6 +26,7 @@ namespace StrawberryShake.Tools
         {
             _hasErrors = true;
             error.Write();
+            _errorReceived();
         }
 
         public void WriteErrors(IEnumerable<HotChocolate.IError> errors)
@@ -32,6 +35,7 @@ namespace StrawberryShake.Tools
             {
                 _hasErrors = true;
                 error.Write();
+                _errorReceived();
             }
         }