+++ src/Core/Types.Tests/Types/InputObjectTypeTests.cs
@@ -1068,7 +1068,7 @@ namespace HotChocolate.Types
 
             // act
             IExecutionResult result = await executor.ExecuteAsync(
-                "{ do(input: { bar: \"abc\" baz: \"def\" }) { bar baz } }");
+                "{ do(input: { bar: \"abc\" baz: \"def\" qux: \"ghi\" }) { bar baz qux } }");
 
             // assert
             result.MatchSnapshot();
@@ -1209,6 +1209,7 @@ namespace HotChocolate.Types
 
             public string Bar { get; }
             public string Baz { get; set; }
+            public string Qux { get; private set; }
         }
     }
 }