+++ src/Core/Types/Types/Introspection/IntrospectionTypeAttribute.cs
@@ -3,7 +3,7 @@
 namespace HotChocolate.Types
 {
     /// <summary>
-    /// Defines that the annotated type is an internal intospection type.
+    /// Defines that the annotated type is an internal introspection type.
     /// </summary>
     [AttributeUsage(
         AttributeTargets.Class,