+++ src/Templates/StarWars/content/StarWars/Characters/Unit.cs
@@ -1,0 +1,11 @@
+namespace StarWars.Characters
+{
+    /// <summary>
+    /// Different units of measurement.
+    /// </summary>
+    public enum Unit
+    {
+        Foot,
+        Meters
+    }
+}